agent/pyproject.toml
[project]
name = "lost-london-agent"
version = "0.1.0"
description = "VIC - Lost London voice assistant powered by Pydantic AI"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai>=0.0.40",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "asyncpg>=0.29.0",
    "groq>=0.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
